+++ api/discovery/_metadata.js
@@ -35,7 +35,7 @@ const DISCOVERY_GAME_SLUGS = Object.freeze([
   "kropkidots", "pearlloop", "starbattle", "parcelpatch", "dominomosaic", "windbow",
   "knightstour", "neondivide", "lureline", "aquariumlogic", "morrismeadow", "reboundrelay",
   "meteorminer", "rippleshepherd", "pollenpatrol", "starwheel", "firebreakcommand", "sumshade",
-  "lanternwake", "acornascent",
+  "lanternwake", "acornascent", "moonscale",
 ]);
 
 const CURATED_TRENDING_SLUGS = Object.freeze([